+++ include/circt-c/Dialect/ESI.h
@@ -18,7 +18,7 @@ extern "C" {
 #endif
 
 MLIR_DECLARE_CAPI_DIALECT_REGISTRATION(ESI, esi);
-void registerESIPasses();
+MLIR_CAPI_EXPORTED void registerESIPasses();
 
 #ifdef __cplusplus
 }
