+++ lib/Bindings/Python/CIRCTModule.cpp
@@ -24,37 +24,6 @@ namespace py = pybind11;
 
 static void registerPasses() { registerSVPasses(); }
 
-/// Taken from PybindUtils.h in MLIR
-/// Accumulates into a python file-like object, either writing text (default)
-/// or binary.
-class PyFileAccumulator {
-public:
-  PyFileAccumulator(pybind11::object fileObject, bool binary)
-      : pyWriteFunction(fileObject.attr("write")), binary(binary) {}
-
-  void *getUserData() { return this; }
-
-  MlirStringCallback getCallback() {
-    return [](MlirStringRef part, void *userData) {
-      pybind11::gil_scoped_acquire();
-      PyFileAccumulator *accum = static_cast<PyFileAccumulator *>(userData);
-      if (accum->binary) {
-        // Note: Still has to copy and not avoidable with this API.
-        pybind11::bytes pyBytes(part.data, part.length);
-        accum->pyWriteFunction(pyBytes);
-      } else {
-        pybind11::str pyStr(part.data,
-                            part.length); // Decodes as UTF-8 by default.
-        accum->pyWriteFunction(pyStr);
-      }
-    };
-  }
-
-private:
-  pybind11::object pyWriteFunction;
-  bool binary;
-};
-
 PYBIND11_MODULE(_circt, m) {
   m.doc() = "CIRCT Python Native Extension";
   registerPasses();
@@ -87,7 +56,7 @@ PYBIND11_MODULE(_circt, m) {
       "Register CIRCT dialects on a PyMlirContext.");
 
   m.def("export_verilog", [](MlirModule mod, py::object fileObject) {
-    PyFileAccumulator accum(fileObject, false);
+    circt::python::PyFileAccumulator accum(fileObject, false);
     py::gil_scoped_release();
     mlirExportVerilog(mod, accum.getCallback(), accum.getUserData());
   });
