+++ lib/CAPI/Dialect/SV.cpp
@@ -10,7 +10,6 @@
 #include "mlir/CAPI/IR.h"
 #include "mlir/CAPI/Registration.h"
 #include "mlir/CAPI/Support.h"
-#include "mlir/Pass/PassRegistry.h"
 
 MLIR_DEFINE_CAPI_DIALECT_REGISTRATION(SystemVerilog, sv, circt::sv::SVDialect)
